:truck: Move: 폴더구조 변경
- dummy-data 폴더 추가
- 캐러셀 components폴더로 이동
