+++ src/components/home/HeroCarousel.tsx
